#1940 - Initial migration of Compass from TestApp to SDK
