+++ android/java/MapboxGLAndroidSDK/src/main/res/drawable-mdpi/compass.png
