+++ android/java/MapboxGLAndroidSDK/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="compassContentDescription">Map compass. Click to reset the map rotation to North.</string>
+</resources>
